コメントに評価値を作る　draft_listがログイン中のユーザと一致しないものを見れない仕様に変更

+++ blog/views.py
@@ -5,10 +5,14 @@ from django.shortcuts import redirect
 from django.contrib.auth.decorators import login_required
 from .forms import PostForm, CommentForm
 from .models import Post, Comment
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import login, authenticate
+
+from django.views.generic.edit import FormView
 # Create your views here.
 
 def post_list(request):
-    posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('published_date')
+    posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('-published_date')
     return render(request, 'blog/post_list.html', {'posts':posts})
 
 def post_detail(request, pk):
@@ -31,6 +35,8 @@ def post_new(request):
 @login_required
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if post.author != request.user:
+        return redirect('post_detail', pk=pk)
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
@@ -44,7 +50,11 @@ def post_edit(request, pk):
 
 @login_required
 def post_draft_list(request):
-    posts=Post.objects.filter(published_date__isnull=True).order_by('created_date')
+    posts=Post.objects.filter(
+        published_date__isnull=True
+    ).filter(
+        author=request.user
+    ).order_by('created_date')
     return render(request, 'blog/post_draft_list.html', {'posts': posts})
 
 @login_required
@@ -56,22 +66,27 @@ def post_publish(request, pk):
 @login_required
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
+    if post.author == request.user:
+        post.delete()
     return redirect('post_list')
 
 def add_comment_to_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
     if request.method == "POST":
         form = CommentForm(request.POST)
-        if form.is_valid:
-            comment = form.save(commit=False)
-            comment.post = post
-            comment.save()
-            return redirect('post_detail', pk=post.pk)
+        if int(request.POST['evaluation_value']) >= 0 and int(request.POST['evaluation_value']) <= 5:
+            if form.is_valid:
+                comment = form.save(commit=False)
+                comment.post = post
+                comment.save()
+                return redirect('post_detail', pk=post.pk)
+        else:
+            form.errors['evaluation_value'] = '0~5の範囲で指定してください'
     else:
         form = CommentForm()
     
     return render(request, 'blog/add_comment_to_post.html', {'form': form})
+
 
 @login_required
 def comment_approve(request, pk):
@@ -85,3 +100,16 @@ def comment_remove(request, pk):
     comment.delete()
     return redirect('post_detail', pk=comment.post.pk)
 
+def signup(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data.get('username')
+            raw_password = form.cleaned_data.get('password1')
+            user = authenticate(username=username, password=raw_password)
+            login(request, user)
+            return redirect('post_list')
+    else:
+        form = UserCreationForm()
+    return render(request, 'blog/signup.html', {'form': form})
